add one type definition
